code/backend/requirements.txt
Django==2.2.10
django-cors-headers==3.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
